refactor: eatinorders 패키지 분리
