+++ src/main/java/kitchenpos/eatinorders/domain/JpaOrderTableRepository.java
@@ -1,8 +1,8 @@
-package kitchenpos.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
+package kitchenpos.eatinorders.domain;
 
 import java.util.UUID;
+
+import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaOrderTableRepository extends OrderTableRepository, JpaRepository<OrderTable, UUID> {
 }
