+++ src/main/java/kitchenpos/eatinorders/domain/OrderTable.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.eatinorders.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
