+++ src/main/java/kitchenpos/eatinorders/domain/OrderType.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.eatinorders.domain;
 
 public enum OrderType {
     DELIVERY, TAKEOUT, EAT_IN
